Refactor: 프로필 설정 hook 분리
