+++ src/components/form/formLayout.tsx
@@ -4,7 +4,7 @@ import { FieldValues, FormProvider, UseFormReturn } from "react-hook-form";
 
 interface FormLayoutProps<T extends FieldValues> {
   methods: UseFormReturn<T>;
-  onSubmit: (data: T) => void;
+  onSubmit?: (data: T) => void;
   children: ReactNode;
   error?: string | null;
 }
@@ -17,7 +17,7 @@ export default function FormLayout<T extends FieldValues>({
   return (
     <FormProvider {...methods}>
       <form
-        onSubmit={methods.handleSubmit(onSubmit)}
+        onSubmit={onSubmit ? methods.handleSubmit(onSubmit) : undefined}
         className="flex flex-col gap-3 w-full max-w-md p-12 justify-center"
       >
         {children}
